test: update tests after unroll_sequences move
